+++ basteh/src/test_utils.rs
@@ -3,48 +3,72 @@ use std::{cmp::Ordering, collections::HashSet, time::Duration};
 use crate::{dev::*, *};
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-////////////////////////////////////////////////////    Store tests     ////////////////////////////////////////////////////
+////////////////////////////////////////////////////    Basteh tests     ////////////////////////////////////////////////////
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-pub async fn test_store<S>(store: S)
+pub async fn test_store<P>(store: P)
 where
-    S: 'static + Store,
+    P: 'static + Provider,
 {
-    let storage = Storage::build().store(store).no_expiry().finish();
+    let store = Basteh::build().provider(store).finish();
 
     tokio::join!(
-        test_store_methods(storage.clone()),
-        test_store_keys(storage.clone())
+        test_store_methods(store.clone()),
+        test_store_bytes(store.clone()),
+        test_store_numbers(store.clone()),
+        test_store_keys(store.clone())
     );
 }
 
-pub async fn test_store_methods(storage: Storage) {
+pub async fn test_store_methods(store: Basteh) {
     let key = "store_key";
     let value = "val";
 
-    assert!(storage.set(key, value).await.is_ok());
+    assert!(store.set(key, value).await.is_ok());
 
-    let get_res = storage.get::<String>(key).await;
+    let get_res = store.get::<String>(key).await;
     assert!(get_res.is_ok());
     assert_eq!(get_res.unwrap(), Some(value.to_owned()));
 
-    let contains_res = storage.contains_key(key).await;
+    let contains_res = store.contains_key(key).await;
     assert!(contains_res.is_ok());
     assert!(contains_res.unwrap());
 
-    assert!(storage.delete(key).await.is_ok());
+    assert!(store.delete(key).await.is_ok());
 
-    let get_res = storage.get::<String>(key).await;
+    let get_res = store.get::<String>(key).await;
     assert!(get_res.is_ok());
     assert!(get_res.unwrap().is_none());
 
-    let contains_res = storage.contains_key(key).await;
+    let contains_res = store.contains_key(key).await;
     assert!(contains_res.is_ok());
     assert!(!contains_res.unwrap());
 }
 
-pub async fn test_store_keys(storage: Storage) {
-    let storage = storage.scope("TEST_SCOPE");
+pub async fn test_store_numbers(store: Basteh) {
+    let key = "number_key";
+    let value = 1337;
+
+    assert!(store.set(key, value).await.is_ok());
+
+    let get_res = store.get::<i64>(key).await;
+    assert!(get_res.is_ok());
+    assert_eq!(get_res.unwrap(), Some(value));
+}
+
+pub async fn test_store_bytes(store: Basteh) {
+    let key = "bytes_key";
+    let value = b"Some Value";
+
+    assert!(store.set(key, value.as_ref()).await.is_ok());
+
+    let get_res = store.get::<Vec<u8>>(key).await;
+    assert!(get_res.is_ok());
+    assert_eq!(get_res.unwrap(), Some(value.to_vec()));
+}
+
+pub async fn test_store_keys(store: Basteh) {
+    let store = store.scope("TEST_SCOPE");
     let value = "val";
 
     let mut keys = HashSet::new();
@@ -56,9 +80,9 @@ pub async fn test_store_keys(storage: Storage) {
     keys.insert(String::from("key6"));
 
     for key in keys.iter() {
-        assert!(storage.set(&key, value).await.is_ok());
+        assert!(store.set(&key, value).await.is_ok());
     }
-    let retrieved_keys = storage
+    let retrieved_keys = store
         .keys()
         .await
         .unwrap()
@@ -69,28 +93,28 @@ pub async fn test_store_keys(storage: Storage) {
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-////////////////////////////////////////////////////    Expiry tests     ///////////////////////////////////////////////////
+//////////////////////////////////////////////////    Expiration tests     /////////////////////////////////////////////////
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 /// Testing the expiry basics by setting a key and expiry, then waiting for it to expire
-pub async fn test_expiry_basics(storage: Storage, delay_secs: u64) {
+pub async fn test_expiry_basics(store: Basteh, delay_secs: u64) {
     let delay = Duration::from_secs(delay_secs);
     let key = "expiring_key";
     let value = "val";
 
-    // Expiry for invalid key should return None
-    assert!(storage.expiry(key).await.unwrap().is_none());
+    // Provider for invalid key should return None
+    assert!(store.expiry(key).await.unwrap().is_none());
 
     // Testing set and get before expiry
-    assert!(storage.set(key, value).await.is_ok());
-    assert!(storage.expire(key, delay).await.is_ok());
+    assert!(store.set(key, value).await.is_ok());
+    assert!(store.expire(key, delay).await.is_ok());
     assert_eq!(
-        storage.get::<String>(key).await.unwrap(),
+        store.get::<String>(key).await.unwrap(),
         Some(value.to_owned())
     );
 
     // The exact duration depends on the implementation
-    let exp = storage.expiry(key).await.unwrap().unwrap();
+    let exp = store.expiry(key).await.unwrap().unwrap();
     assert!(exp.as_secs() > 0);
     assert!(exp.as_secs() <= delay_secs);
 
@@ -98,25 +122,25 @@ pub async fn test_expiry_basics(storage: Storage, delay_secs: u64) {
     tokio::time::sleep(Duration::from_secs(delay_secs + 1)).await;
 
     // Check if extended item has been expired
-    assert_eq!(storage.get::<String>(key).await.unwrap(), None);
+    assert_eq!(store.get::<String>(key).await.unwrap(), None);
 }
 
 /// Testing extending functionality by setting an expiry and extending it later,
 /// The key shouldn't be expired before the sum of default expiry and extended time
-pub async fn test_expiry_extend(storage: Storage, delay_secs: u64) {
+pub async fn test_expiry_extend(store: Basteh, delay_secs: u64) {
     let delay = Duration::from_secs(delay_secs);
     let key = "extended_expiring_key";
     let value = "val";
 
     // Testing set and get before expiry
-    assert!(storage.set(key, value).await.is_ok());
-    assert!(storage.expire(key, delay).await.is_ok());
+    assert!(store.set(key, value).await.is_ok());
+    assert!(store.expire(key, delay).await.is_ok());
 
     // Testing extending the expiry time
     // The exact duration depends on the implementation
-    storage.extend(key, delay).await.unwrap();
+    store.extend(key, delay).await.unwrap();
 
-    let exp = storage.expiry(key).await.unwrap().unwrap();
+    let exp = store.expiry(key).await.unwrap().unwrap();
     assert!(exp.as_secs() >= delay_secs);
     assert!(exp.as_secs() <= delay_secs * 2);
 
@@ -125,7 +149,7 @@ pub async fn test_expiry_extend(storage: Storage, delay_secs: u64) {
 
     // Check if extended item still exist
     assert_eq!(
-        storage.get::<String>(key).await.unwrap(),
+        store.get::<String>(key).await.unwrap(),
         Some(value.to_owned())
     );
 
@@ -133,100 +157,100 @@ pub async fn test_expiry_extend(storage: Storage, delay_secs: u64) {
     tokio::time::sleep(Duration::from_secs(delay_secs + 1)).await;
 
     // Check if extended item has been expired
-    assert_eq!(storage.get::<String>(key).await.unwrap(), None);
+    assert_eq!(store.get::<String>(key).await.unwrap(), None);
 }
 
 /// Testing persist, by setting an expiry for a key and making it persistant later
-pub async fn test_expiry_persist(storage: Storage, delay_secs: u64) {
+pub async fn test_expiry_persist(store: Basteh, delay_secs: u64) {
     let delay = Duration::from_secs(delay_secs);
     let key = "persistant_key";
     let value = "val";
 
-    assert!(storage.set(key, value).await.is_ok());
-    assert!(storage.expire(key, delay).await.is_ok());
-    assert!(storage.persist(key).await.is_ok());
+    assert!(store.set(key, value).await.is_ok());
+    assert!(store.expire(key, delay).await.is_ok());
+    assert!(store.persist(key).await.is_ok());
 
     // Adding some error to the delay, for the implementers sake
     tokio::time::sleep(Duration::from_secs(delay_secs + 1)).await;
 
     // Check if persistent key is still there
     assert_eq!(
-        storage.get::<String>(key).await.unwrap(),
+        store.get::<String>(key).await.unwrap(),
         Some(value.to_owned())
     );
 }
 
 /// Testing persist, by setting an expiry for a key and making it persistant later
-pub async fn test_mutate_sould_not_change_expiry(storage: Storage, delay_secs: u64) {
+pub async fn test_mutate_sould_not_change_expiry(store: Basteh, delay_secs: u64) {
     let delay = Duration::from_secs(delay_secs);
     let key = "mutated_key";
     let value = 1200;
 
-    assert!(storage.set(key, value).await.is_ok());
-    assert!(storage.expire(key, delay).await.is_ok());
+    assert!(store.set(key, value).await.is_ok());
+    assert!(store.expire(key, delay).await.is_ok());
 
-    assert!(storage.mutate(key, |m| m.incr(1300)).await.is_ok());
+    assert!(store.mutate(key, |m| m.incr(1300)).await.is_ok());
 
-    assert_eq!(storage.get::<i64>(key).await.unwrap(), Some(2500));
+    assert_eq!(store.get::<i64>(key).await.unwrap(), Some(2500));
 
     // Check if persistent expiry have changed
-    assert!(storage.expiry(key).await.unwrap().is_some());
+    assert!(store.expiry(key).await.unwrap().is_some());
 }
 
 /// Testing if calling set after expire, clears expiration from the key
-pub async fn test_expiry_set_clearing(storage: Storage, delay_secs: u64) {
+pub async fn test_expiry_set_clearing(store: Basteh, delay_secs: u64) {
     let delay = Duration::from_secs(delay_secs);
     let key = "set_after_expire_key";
     let value = "val";
 
-    assert!(storage.set(key, value).await.is_ok());
-    assert!(storage.expire(key, delay).await.is_ok());
-    assert!(storage.set(key, value).await.is_ok());
+    assert!(store.set(key, value).await.is_ok());
+    assert!(store.expire(key, delay).await.is_ok());
+    assert!(store.set(key, value).await.is_ok());
 
     tokio::time::sleep(Duration::from_secs((delay_secs) + 1)).await;
 
     // Check if calling set twice cleaerd the expire
     assert_eq!(
-        storage.get::<String>(key).await.unwrap(),
+        store.get::<String>(key).await.unwrap(),
         Some(value.to_owned())
     );
 }
 
 /// Testing if second call to expire overrides the first one
 /// The key should expire in this test, as we set a shorter period second time
-pub async fn test_expiry_override_shorter(storage: Storage, delay_secs: u64) {
+pub async fn test_expiry_override_shorter(store: Basteh, delay_secs: u64) {
     let delay = Duration::from_secs(delay_secs);
     let key = "expire_override_shorter_key";
     let value = "val";
 
     // Test if second call to expire overwrites the first expiry
-    assert!(storage.set(key, value).await.is_ok());
-    assert!(storage.expire(key, delay * 5).await.is_ok());
-    assert!(storage.expire(key, delay).await.is_ok());
+    assert!(store.set(key, value).await.is_ok());
+    assert!(store.expire(key, delay * 5).await.is_ok());
+    assert!(store.expire(key, delay).await.is_ok());
 
     tokio::time::sleep(Duration::from_secs(delay_secs + 1)).await;
 
     // Check if calling set twice cleaerd the expire
-    assert_eq!(storage.get::<String>(key).await.unwrap(), None);
+    assert_eq!(store.get::<String>(key).await.unwrap(), None);
 }
 
 /// Testing if second call to expire overrides the first one
 /// The key shouldn't expire in this test, as we set a longer period second time
-pub async fn test_expiry_override_longer(storage: Storage, delay_secs: u64) {
+pub async fn test_expiry_override_longer(store: Basteh, delay_secs: u64) {
     let delay = Duration::from_secs(delay_secs);
     let key = "expire_override_longer_key";
     let value = "val";
 
     // Test if second call to expire overwrites the first expiry
-    assert!(storage.set(key, value).await.is_ok());
-    assert!(storage.expire(key, delay).await.is_ok());
-    assert!(storage.expire(key, delay * 5).await.is_ok());
+    assert!(store.set(key, value).await.is_ok());
+    assert!(store.expire(key, delay).await.is_ok());
+    assert!(store.expire(key, delay * 5).await.is_ok());
 
     tokio::time::sleep(Duration::from_secs(delay_secs + 1)).await;
 
     // Check if calling set twice cleaerd the expire
     assert_eq!(
-        storage.get::<String>(key).await.unwrap(),
+        store.get::<String>(key).await.unwrap(),
         Some(value.to_owned())
     );
 }
@@ -234,37 +258,36 @@ pub async fn test_expiry_override_longer(storage: Storage, delay_secs: u64) {
 // delay_secs is the duration of time we set for expiry and we wait to see
 // the result, it should depend on how much delay an implementer has between
 // getting a command and executing it
-pub async fn test_expiry<S, E>(store: S, expiry: E, delay_secs: u64)
+pub async fn test_expiry<P>(provider: P, delay_secs: u64)
 where
-    S: 'static + Store,
-    E: 'static + Expiry,
+    P: 'static + Provider,
 {
-    let storage = Storage::build().store(store).expiry(expiry).finish();
+    let store = Basteh::build().provider(provider).finish();
 
     tokio::join!(
-        test_expiry_basics(storage.clone(), delay_secs),
-        test_mutate_sould_not_change_expiry(storage.clone(), delay_secs,),
-        test_expiry_extend(storage.clone(), delay_secs),
-        test_expiry_persist(storage.clone(), delay_secs),
-        test_expiry_set_clearing(storage.clone(), delay_secs),
-        test_expiry_override_shorter(storage.clone(), delay_secs),
-        test_expiry_override_longer(storage, delay_secs)
+        test_expiry_basics(store.clone(), delay_secs),
+        test_mutate_sould_not_change_expiry(store.clone(), delay_secs,),
+        test_expiry_extend(store.clone(), delay_secs),
+        test_expiry_persist(store.clone(), delay_secs),
+        test_expiry_set_clearing(store.clone(), delay_secs),
+        test_expiry_override_shorter(store.clone(), delay_secs),
+        test_expiry_override_longer(store, delay_secs)
     );
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////    Expiry Store tests     ////////////////////////////////////////////////
+///////////////////////////////////////////////    Basteh-Expiration tests     //////////////////////////////////////////////
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-pub async fn test_expiry_store_basics(storage: Storage, delay_secs: u64) {
+pub async fn test_expiry_store_basics(store: Basteh, delay_secs: u64) {
     let delay = Duration::from_secs(delay_secs);
     let key = "expiry_store_key";
     let value = "value";
 
     // Test set and get expiring
-    assert!(storage.set_expiring(key, value, delay).await.is_ok());
+    assert!(store.set_expiring(key, value, delay).await.is_ok());
 
-    let (v, e) = storage.get_expiring::<String>(key).await.unwrap().unwrap();
+    let (v, e) = store.get_expiring::<String>(key).await.unwrap().unwrap();
     assert_eq!(&v, &value);
     assert!(e.unwrap().as_secs() > 0);
     assert!(e.unwrap().as_secs() <= delay_secs);
@@ -273,20 +296,20 @@ pub async fn test_expiry_store_basics(storage: Storage, delay_secs: u64) {
     tokio::time::sleep(Duration::from_secs(delay_secs + 1)).await;
 
     // Check if first item expired as expected
-    assert_eq!(storage.get_expiring::<String>(key).await.unwrap(), None);
+    assert_eq!(store.get_expiring::<String>(key).await.unwrap(), None);
 }
 
 /// Testing if second call to expire overrides the first one
 /// The key should expire in this test, as we set a shorter period second time
-pub async fn test_expiry_store_override_shorter(storage: Storage, delay_secs: u64) {
+pub async fn test_expiry_store_override_shorter(store: Basteh, delay_secs: u64) {
     let delay = Duration::from_secs(delay_secs);
     let key = "expire_store_override_shorter_key";
     let value = "value";
 
     // Test if second call to set expiring, overwrites expiry for key
-    assert!(storage.set_expiring(key, value, delay).await.is_ok());
-    assert!(storage.set_expiring(key, value, delay * 2).await.is_ok());
-    let exp = storage.expiry(key).await.unwrap().unwrap();
+    assert!(store.set_expiring(key, value, delay).await.is_ok());
+    assert!(store.set_expiring(key, value, delay * 2).await.is_ok());
+    let exp = store.expiry(key).await.unwrap().unwrap();
     assert!(exp.as_secs() > delay_secs);
     assert!(exp.as_secs() <= delay_secs * 2);
 
@@ -295,22 +318,22 @@ pub async fn test_expiry_store_override_shorter(storage: Storage, delay_secs: u6
 
     // Check if the second call to set overwrites expiry
     assert_eq!(
-        storage.get::<String>(key).await.unwrap(),
+        store.get::<String>(key).await.unwrap(),
         Some(value.to_owned())
     );
 }
 
 /// Testing if second call to expire overrides the first one
 /// The key shouldn't expire in this test, as we set a longer period second time
-pub async fn test_expiry_store_override_longer(storage: Storage, delay_secs: u64) {
+pub async fn test_expiry_store_override_longer(store: Basteh, delay_secs: u64) {
     let delay = Duration::from_secs(delay_secs);
     let key = "expire_store_override_longer_key";
     let value = "value";
 
     // Test if second call to set expiring, overwrites expiry for key
-    assert!(storage.set_expiring(key, value, delay * 2).await.is_ok());
-    assert!(storage.set_expiring(key, value, delay).await.is_ok());
-    let exp = storage.expiry(key).await.unwrap().unwrap();
+    assert!(store.set_expiring(key, value, delay * 2).await.is_ok());
+    assert!(store.set_expiring(key, value, delay).await.is_ok());
+    let exp = store.expiry(key).await.unwrap().unwrap();
     assert!(exp.as_secs() > 0);
     assert!(exp.as_secs() <= delay_secs);
 
@@ -318,126 +341,105 @@ pub async fn test_expiry_store_override_longer(storage: Storage, delay_secs: u64
     tokio::time::sleep(Duration::from_secs(delay_secs + 1)).await;
 
     // Check if the second call to set overwrites expiry
-    assert_eq!(storage.get::<String>(key).await.unwrap(), None);
+    assert_eq!(store.get::<String>(key).await.unwrap(), None);
 }
 
 /// Testing if mutation after expiry works as expected
-pub async fn test_expiry_store_mutate_after_expiry(storage: Storage, delay_secs: u64) {
+pub async fn test_expiry_store_mutate_after_expiry(store: Basteh, delay_secs: u64) {
     let delay = Duration::from_secs(delay_secs);
     let key = "expire_store_mutate_after_expiry_key";
     let value = 1000;
 
     // Set a number and set expiry for it
-    assert!(storage.set(key, value).await.is_ok());
-    assert!(storage.expire(key, delay).await.is_ok());
+    assert!(store.set(key, value).await.is_ok());
+    assert!(store.expire(key, delay).await.is_ok());
 
     // Adding some error to the delay, for the implementers sake
     tokio::time::sleep(Duration::from_secs(delay_secs + 1)).await;
 
     // Mutate the number
-    storage.mutate(key, |m| m.incr(100)).await.unwrap();
-    assert_eq!(storage.get::<i64>(key).await.unwrap(), Some(100))
+    store.mutate(key, |m| m.incr(100)).await.unwrap();
+    assert_eq!(store.get::<i64>(key).await.unwrap(), Some(100))
 }
 
 // delay_secs is the duration of time we set for expiry and we wait to see
 // the result, it should depend on how much delay an implementer has between
 // getting a command and executing it
-pub async fn test_expiry_store<S>(store: S, delay_secs: u64)
+pub async fn test_expiry_store<P>(provider: P, delay_secs: u64)
 where
-    S: 'static + ExpiryStore,
+    P: 'static + Provider,
 {
-    let storage = Storage::build().store(store).finish();
+    let store = Basteh::build().provider(provider).finish();
 
     tokio::join!(
-        test_expiry_store_basics(storage.clone(), delay_secs),
-        test_expiry_store_override_shorter(storage.clone(), delay_secs),
-        test_expiry_store_override_longer(storage.clone(), delay_secs),
-        test_expiry_store_mutate_after_expiry(storage, delay_secs),
+        test_expiry_store_basics(store.clone(), delay_secs),
+        test_expiry_store_override_shorter(store.clone(), delay_secs),
+        test_expiry_store_override_longer(store.clone(), delay_secs),
+        test_expiry_store_mutate_after_expiry(store, delay_secs),
     );
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////    Numbers and Mutation tests     ////////////////////////////////////////////
+///////////////////////////////////////////////////    Mutation tests     //////////////////////////////////////////////////
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-pub async fn test_store_numbers<S>(store: S)
-where
-    S: 'static + Store,
-{
-    let storage = Storage::build().store(store).no_expiry().finish();
-
-    let key = "number_key";
-    let value = 1337;
-
-    assert!(storage.set(key, value).await.is_ok());
-
-    let get_res = storage.get::<i64>(key).await;
-    assert!(get_res.is_ok());
-    assert_eq!(get_res.unwrap(), Some(value));
-}
-
-pub async fn test_mutate_numbers<S>(store: S)
-where
-    S: 'static + Store,
-{
-    let storage = Storage::build().store(store).no_expiry().finish();
-
+pub async fn test_mutate_numbers(store: Basteh) {
     let key = "mutate_number_key";
 
     // Increase by 1600, key doesn't exist so it should be considered 0
-    storage.mutate(key, |m| m.incr(1600)).await.ok();
+    store.mutate(key, |m| m.incr(1600)).await.ok();
 
-    let get_res = storage.get::<i64>(key).await;
+    let get_res = store.get::<i64>(key).await;
     assert!(get_res.is_ok());
     assert_eq!(get_res.unwrap(), Some(1600));
 
     // Decrease by 200
-    storage.mutate(key, |m| m.decr(200)).await.ok();
+    store.mutate(key, |m| m.decr(200)).await.ok();
 
-    let get_res = storage.get::<i64>(key).await;
+    let get_res = store.get::<i64>(key).await;
     assert!(get_res.is_ok());
     assert_eq!(get_res.unwrap(), Some(1400));
 
     // Mutiply by 2
-    storage.mutate(key, |m| m.mul(2)).await.ok();
+    store.mutate(key, |m| m.mul(2)).await.ok();
 
-    let get_res = storage.get::<i64>(key).await;
+    let get_res = store.get::<i64>(key).await;
     assert!(get_res.is_ok());
     assert_eq!(get_res.unwrap(), Some(2800));
 
     // Divide by 4
-    storage.mutate(key, |m| m.div(4)).await.ok();
+    store.mutate(key, |m| m.div(4)).await.ok();
 
-    let get_res = storage.get::<i64>(key).await;
+    let get_res = store.get::<i64>(key).await;
     assert!(get_res.is_ok());
     assert_eq!(get_res.unwrap(), Some(700));
 
     // Set to 100
-    storage.mutate(key, |m| m.set(100)).await.ok();
+    store.mutate(key, |m| m.set(100)).await.ok();
 
-    let get_res = storage.get::<i64>(key).await;
+    let get_res = store.get::<i64>(key).await;
     assert!(get_res.is_ok());
     assert_eq!(get_res.unwrap(), Some(100));
 
     // Conditional if
-    storage
+    store
         .mutate(key, |m| m.if_(Ordering::Equal, 100, |m| m.set(200)))
         .await
         .ok();
 
-    let get_res = storage.get::<i64>(key).await;
+    let get_res = store.get::<i64>(key).await;
     assert!(get_res.is_ok());
     assert_eq!(get_res.unwrap(), Some(200));
 
     // Conditional if else
-    storage
+    store
         .mutate(key, |m| {
             m.if_else(Ordering::Greater, 200, |m| m.decr(100), |m| m.decr(50))
         })
         .await
         .ok();
 
-    let get_res = storage.get::<i64>(key).await;
+    let get_res = store.get::<i64>(key).await;
     assert!(get_res.is_ok());
     assert_eq!(get_res.unwrap(), Some(150));
 
@@ -449,16 +451,28 @@ where
             })
         })
     };
-    storage.mutate(key, mutation).await.ok();
+    store.mutate(key, mutation).await.ok();
 
-    let get_res = storage.get::<i64>(key).await;
+    let get_res = store.get::<i64>(key).await;
     assert!(get_res.is_ok());
     assert_eq!(get_res.unwrap(), Some(175));
 
     // Multi level conditionals
-    storage.mutate(key, mutation).await.ok();
+    store.mutate(key, mutation).await.ok();
 
-    let get_res = storage.get::<i64>(key).await;
+    let get_res = store.get::<i64>(key).await;
     assert!(get_res.is_ok());
     assert_eq!(get_res.unwrap(), Some(125));
+}
+
+// delay_secs is the duration of time we set for expiry and we wait to see
+// the result, it should depend on how much delay an implementer has between
+// getting a command and executing it
+pub async fn test_mutations<P>(provider: P)
+where
+    P: 'static + Provider,
+{
+    let store = Basteh::build().provider(provider).finish();
+
+    tokio::join!(test_mutate_numbers(store.clone()));
 }